pyproject.toml
[tool.poetry]
name = "polls"
version = "0.9.0"
description = "Polls API"
authors = ["gixproject <viacheslavlab@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
psycopg2 = "^2.8.5"
psycopg2-binary = "^2.8.5"
Django = "3.2.3"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.3"
drf-yasg = "^1.20.0"
environs = "^9.2.0"
Werkzeug = "^1.0.1"
isort = "^5.8.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"
flake8 = "^3.8.4"
flake8-print = "^4.0.0"
flake8-quotes = "^3.2.0"
flake8-eradicate = "^1.0.0"
pep8-naming = "^0.11.1"
flake8-functions = "^0.0.4"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
